pyproject.toml
[project]
name = "acp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "acp-sdk>=1.0.0",
    "crewai>=0.141.0",
    "crewai-tools>=0.51.1",
    "duckduckgo-search>=8.1.1",
    "google-generativeai>=0.8.5",
    "load-dotenv>=0.1.0",
    "markdownify>=1.1.0",
    "mcp>=1.11.0",
    "nest-asyncio>=1.6.0",
    "requests>=2.32.4",
    "smolagents[mcp]>=1.20.0",
]
